Rewrote run-mirai script to use KLEE instead
This is a bit of a hack to invoke rustc and KLEE on the file provided.
In the future, it may be worth making this more flexible so that,
for example, it supports multiple files or can be provided with a list
of imported crates.
But this is enough for what I need just now.